backend/requirements.txt
asgiref==3.10.0
astroid==3.3.11
black==25.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
dill==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
isort==6.1.0
joblib==1.5.2
mccabe==0.7.0f
mypy_extensions==1.1.0
mysqlclient==2.2.7
dj-database-url==1.2.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.4.0
pycparser==2.23
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tomlkit==0.13.3
tzdata==2025.2
requests==2.34.0
huggingface_hub>=0.16.0
